requirements.txt
asgiref==3.7.1
Django==4.2.1
django-ckeditor==6.5.1
django-decouple==2.1
django-js-asset==2.0.0
Pillow==9.5.0
sqlparse==0.4.4
